DRY Baysian tests using modules
